// src/modules/riffles/enums/raffle-messages.enum.ts
/* src/modules/raffles/enums/raffle-messages.enum.ts */
export enum RaffleMessages {
  RAFFLE_CREATED = 'Rifa creada exitosamente',
  RAFFLE_NOT_FOUND = 'Rifa no encontrada',
  RAFFLES_LISTED = 'Listado de rifas',
  RAFFLE_UPDATED = 'Rifa actualizada exitosamente',
  RAFFLE_DELETED = 'Rifa eliminada exitosamente',
  PARTICIPATION_ADDED = 'Participación añadida',
  INVALID_ID = 'ID de rifa inválido',
}