// src/modules/riffles/enums/raffle-operation-summaries.enum.ts
/* src/modules/raffles/enums/raffle-operation-summaries.enum.ts */
export enum RaffleOperationSummaries {
  CREATE = 'Crear nueva rifa',
  FIND_ALL = 'Obtener todas las rifas',
  FIND_ONE = 'Obtener rifa por ID',
  UPDATE = 'Actualizar rifa',
  DELETE = 'Eliminar rifa',
  ADD_PARTICIPANT = 'Agregar participante a rifa',
}